+++ curl_cffi/requests/impersonate.py
@@ -29,6 +29,10 @@ BrowserTypeLiteral = Literal[
     "chrome142",
     "chrome145",
     "chrome146",
+    "chrome150",
+    "chrome150_linux",
+    "chrome150_windows",
+    "chrome150_macos",
     "chrome99_android",
     "chrome131_android",
     # Safari
@@ -79,7 +83,7 @@ BrowserTypeLiteral = Literal[
 ]
 
 
-DEFAULT_CHROME = "chrome146"
+DEFAULT_CHROME = "chrome150"
 DEFAULT_EDGE = "edge101"
 DEFAULT_SAFARI = "safari2601"
 DEFAULT_SAFARI_IOS = "safari260_ios"
@@ -91,7 +95,7 @@ DEFAULT_TOR = "tor145"
 
 
 REAL_TARGET_MAP = {
-    "chrome": "chrome146",
+    "chrome": "chrome150",
     "edge": "edge101",
     "safari": "safari2601",
     "safari_ios": "safari260_ios",
@@ -146,6 +150,10 @@ class BrowserType(str, Enum):  # TODO: remove in version 1.x
     chrome142 = "chrome142"
     chrome145 = "chrome145"
     chrome146 = "chrome146"
+    chrome150 = "chrome150"
+    chrome150_linux = "chrome150_linux"
+    chrome150_windows = "chrome150_windows"
+    chrome150_macos = "chrome150_macos"
     chrome99_android = "chrome99_android"
     chrome131_android = "chrome131_android"
     safari153 = "safari153"