+++ scripts/bump_version.sh
@@ -4,7 +4,7 @@ VERSION=$1
 UPSTREAM_VERSION=$2
 
 # Makefile
-gsed "s/^VERSION := .*/VERSION := ${UPSTREAM_VERSION}/g" -i Makefile
+gsed "s/^VERSION ?= .*/VERSION ?= ${UPSTREAM_VERSION}/g" -i Makefile
 
 # pyproject.toml
 gsed "s/^version = .*/version = \"${VERSION}\"/g" -i pyproject.toml